terminar ventana de creacion de tipo de boleto

+++ src/SistemaZoologico.Dominio/Aplicacion/VentaBoletos/Commandos/CrearTipoBoleto.cs
@@ -2,13 +2,13 @@
 {
     public class CrearTipoBoleto
     {
-        public CrearTipoBoleto(string descripcion, double precio)
+        public CrearTipoBoleto(string descripcion, float precio)
         {
             Descripcion = descripcion;
             Precio = precio;
         }
 
         public string Descripcion { get; set; }
-        public double Precio { get; set; }
+        public float Precio { get; set; }
     }
 }
